+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 # Ocaml-jetbrains Changelog
 
 ## [Unreleased]
+
+### Added
+
+- Dune Build, Dune Exec, and Dune Test run configurations with OPAM toolchain support.